requirements.txt
Django
onetimepass
autobahn
nose
websocket-client
pyqrcode
pyyaml 
